requirements.txt
fire
tqdm
numpy
ujson
pandas
dotmap
sklearn
pytorch